ve/fs: Use comparison of signed integers in ve_mount_allowed()
ve->mnr_nr is a signed integer which may become negative in some cases:
* if the increments and decrements of ve->mnr_nr race against each other
  and the decrements win;
* if something is mounted by one VE but unmounted by another VE.

sysctl_ve_mount_nr is unsigned.

So the comparison (ve->mnr_nr < sysctl_ve_mount_nr) can actually be
compiled as ((unsigned int)ve->mnr_nr < sysctl_ve_mount_nr).
ve_mount_allowed() would return 0 in that case and the mount operation
would fail as a result.

This patch fixes the problem.

https://jira.sw.ru/browse/PSBM-42825

+++ include/linux/ve.h
@@ -238,7 +238,7 @@ static inline int ve_mount_allowed(void)
 {
 	struct ve_struct *ve = get_exec_env();
 
-	return ve_is_super(ve) || ve->mnt_nr < sysctl_ve_mount_nr;
+	return ve_is_super(ve) || ve->mnt_nr < (int)sysctl_ve_mount_nr;
 }
 
 static inline void ve_mount_nr_inc(void)
